try split arm-code and amd-code
